Don't append flattened device tree to kernel
Keeping the FTD seperate allows for overriding the bootargs from
U-Boot as well as having the option to make use of the
multiplatform features to swap in FTDs for other boards.

Update the kernel config file, and the README.  Remove the
duImage-m2.sh script which is not necessary any more.